+++ tests/cli.rs
@@ -166,6 +166,23 @@ fn module_info() {
 }
 
 #[test]
+fn module_info_relannis() {
+    let mut cmd = Command::cargo_bin("annatto").unwrap();
+
+    let output = cmd
+        .env("NO_COLOR", "1")
+        .arg("info")
+        .arg("relannis")
+        .output()
+        .unwrap();
+    cmd.assert().success();
+
+    let output = std::str::from_utf8(&output.stdout).unwrap();
+
+    assert_snapshot!(output);
+}
+
+#[test]
 fn graph_op_info() {
     let mut cmd = Command::cargo_bin("annatto").unwrap();